Refactor code and elaborate README.md
